fix random noise

+++ hw_asr/augmentations/wave_augmentations/RandomNoise.py
@@ -1,4 +1,5 @@
 import numpy as np
+import torch
 from torch import Tensor
 
 from hw_asr.augmentations.base import AugmentationBase
@@ -14,7 +15,7 @@ class RandomNoiseBase(AugmentationBase):
         noise = np.random.normal(size=audio.shape)
         noise /= np.abs(noise).max()  # [-1, 1]
         noised_audio = audio + noise_ampl / np.abs(audio).max() * noise
-        return noised_audio
+        return noised_audio.double()
 
 
 class RandomNoise(AugmentationBase):
